Updated for I2C and Functionality
I think there's still an I2C arbitration bug in this code in the setText function but it seems to work most of the time.  I believe it's related to the I2C arbitration but am not sure as the error is vague and infrequent.  Otherwise seems very reliable and I am able to get accurate readings in various modes of operation (GForce and Acceleration).