Updated for stability and documentation
I still am not sure of the actual purpose of the write to the display address (0x62) for Data Register 0x08 the value of 0xff or 0xaa that happens on lines 62 & 63.  The code is there in the .py (writes 0xaa) and .cpp (writes 0xff) referenced examples but the documentation isn't clear enough to explain it's purpose.  I found the datasheet but nothing specific for these flags that I could decipher from it.  The LCD datasheet (PDF) is here: https://www.sparkfun.com/datasheets/LCD/HD44780.pdf