ref#1: refactored base i2c library
* used static functions / variables to restrict access to source files
-> therefore avoiding unnecessary conflict

* use of more adequate data types

* added better documentation

* more explicit variable and function naming

* better structured code

* removed global variables - fits into "best practices" and prevents
possible conflicts caused by the lack of locks -> users may try to call
the same function while the the bus may be busy doing something else and
things can get dirty this way.

* removed DEFINES -> DEFINES might hide some conflicts and errors. This
might happen when the user sends the DEFINEd literal to a function as a
paramenter. So, solving an "unexplicable" error might lead to entire
days lost in finding the cause.