Merges + bumped NPM version (#366)
* Updated for I2C and Functionality

I think there's still an I2C arbitration bug in this code in the setText function but it seems to work most of the time.  I believe it's related to the I2C arbitration but am not sure as the error is vague and infrequent.  Otherwise seems very reliable and I am able to get accurate readings in various modes of operation (GForce and Acceleration).

* New File to Manage the Grove LCD RGB Display

New addition to manage the Grove LCD display...there is a bug with the setText and the I2C arbitration but mostly works as expected.  I am not sure what I am missing here but advice is appreciated.

* Update package.json

added mathjs dependency

* bumped major version

* Minor Tweak to turnOffDisplauy

Just a tweak i realized a copy paste inconsistency with the resetDisplay() and turnOffDisplay() functions.

* Updated for stability and documentation

I still am not sure of the actual purpose of the write to the display address (0x62) for Data Register 0x08 the value of 0xff or 0xaa that happens on lines 62 & 63.  The code is there in the .py (writes 0xaa) and .cpp (writes 0xff) referenced examples but the documentation isn't clear enough to explain it's purpose.  I found the datasheet but nothing specific for these flags that I could decipher from it.  The LCD datasheet (PDF) is here: https://www.sparkfun.com/datasheets/LCD/HD44780.pdf

* reverted changes made in a contribution who was not backward compatible

* Deleted class who was never used and was not backward compatible

* exporting AccelerationI2cSensor